[WS-04] Pre-build review fixes: no false content errors on placement, two forward decls
An independent review of all 34 changed files found the following.

- PlaceTower asked the traps table first, and UDFContentSubsystem logs a
  missing row as an Error. Placing a lance therefore logged "table 'traps'
  has no row 'lance'", which would have failed about 8 DF.Unit.Tower tests
  and put a false error in every match. Trap, tower and wave-condition ids
  are now checked quietly with Ids(Table).Contains: a bad id from a player
  is a refusal, not broken content.
- DFTowerDamage.h forward-declares AActor, and DFTowerRigComponent.h
  forward-declares UStaticMesh. Both headers compiled only because the
  unity build put another file first.
- The no-wallet test uses AddExpectedMessage(Warning, 1), like the other
  suites.
- Both arms of Heat's conditional are now uint8.

NOT BUILT.

+++ unreal/DeepField/Source/DFTowers/Private/Towers/DFTower.cpp
@@ -407,7 +407,7 @@ void ADFTower::StepWeapon(const FDFTowerRow& Row, float DeltaSeconds)
 			DFBalance::Dial(this, TEXT("beamRampPerSecond"), 0.6f), RampCap);
 		// The cap the ramp is heading for, with the "peak" path, so full heat means "at the cap".
 		const float Cap = RampCap * DFTowerMath::PathFactor(Row, PathLevels, TEXT("peak"));   // BeamRamp's cap
-		Heat = Cap > 1.f ? static_cast<uint8>(FMath::Clamp((Ramp - 1.f) / (Cap - 1.f), 0.f, 1.f) * 255.f + 0.5f) : 0;
+		Heat = Cap > 1.f ? static_cast<uint8>(FMath::Clamp((Ramp - 1.f) / (Cap - 1.f), 0.f, 1.f) * 255.f + 0.5f) : uint8(0);
 		DealDamage(Row, Target, DFTowerMath::EffectiveDamage(Row, PathLevels) * Ramp * DeltaSeconds, DFTags::Damage_Type_Thermal);
 		return;
 	}
